requirements.txt
llama_index
numpy
pandas
openai 
os
python-dotenv